requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.4
openai==1.3.0
requests==2.31.0
gunicorn==21.2.0
